Move device layouts to separate file
